Remove kafkatee.service's ExecStopPost
ExecStopPost is not passed to a shell, so test -f && rm is currently
producing errors. We could write this as rm -f, but on the other hand,
looking at ezd.c, kafkatee seems to perfectly handle fine the case where
a pidfile already exists and only aborts if there is an active lock on
the file, so no reason to do this magic at all. Remove it altogether.

Change-Id: Idc05ca6ad3ee8d6da6d703812dc083badec800c6